gun_control_previous/requirements.txt
pyserial
Flask
flask-socketio
gevent
picamera2
numpy